requirements.txt
fastapi == 0.116.1
uvicorn == 0.35.0
jinja2 == 3.1.6
yt-dlp == 2025.06.30
webvtt-py == 0.5.1
slowapi == 0.1.9
python-multipart
sentence_transformers == 5.0.0
Groq == 0.30.0
pydantic == 2.11.7
chromadb == 1.0.15
python-dotenv == 1.1.1
google-search-results == 2.4.2
serpapi == 0.1.5
langchain == 0.3.27
langchain-groq == 0.3.6
python-jose[cryptography]  # For JWT if needed
